+++ src/main/java/com/coachera/backend/entity/Category.java
@@ -1,14 +1,10 @@
-package com.coachera.backend.entities;
-
-import java.time.LocalDate;
+package com.coachera.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -18,25 +14,13 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "certificates")
+@Table(name="categories")
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class Certificate extends Auditable {
+public class Category extends Auditable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @ManyToOne
-    @JoinColumn(name = "course_id", nullable = false)
-    private Course course;
-
-    @ManyToOne
-    @JoinColumn(name = "student_id", nullable = false)
-    private Student student;
-
     @Column(nullable = false)
-    private LocalDate issuedAt;
-
-    @Column(nullable = false)
-    private String certificateUrl;
+    private String name;
 }
-
