+++ src/main/java/com/coachera/backend/repositories/UserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.coachera.backend.entities.User;
+import com.coachera.backend.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
